Add clickable links to LinkedIn, YouTube and Twitter icons
Implemented clickable links for the LinkedIn , Youtube and Twitter icons. Now, users can navigate directly to the respective profiles by clicking on the associated icons. The 'a' (anchor) HTML element has been utilized for hyperlinking, and the 'target="_blank"' attribute ensures that the links open in a new tab for a seamless user experience.